Introduce LMS Default services
